placePiece() passes all tests once more
Slightly changed the structure of the placements array. We now have a
default array which holds default max piece numbers which will be cloned
into a placements array as each player takes his/her turn and those
values will be decremented instead. Will be reflected in testing in next
commit.